Document how to invoke refactorings in VSCode
Thanks so much for working on this tool! I've wanted automated refactoring support for Elixir for so long!

I've been using VSCode with only Elixir for a long time, so I had to dig around to figure out how to use RefactorEx when I first started trying it out.

This PR adds a paragraph to the README to explain how to access RefactorEx's features.

These are the commands I've found through my research; there may be more, and if so, I'd be happy to add them to this update.